Major refactor - move logic to manager classes, GUI to UI class
